LayoutTests/imported/w3c/web-platform-tests/content-security-policy/style-src-attr-elem/style-src-elem-blocked-attr-allowed-expected.txt

CONSOLE MESSAGE: Unrecognized Content-Security-Policy directive 'style-src-elem'.

CONSOLE MESSAGE: Unrecognized Content-Security-Policy directive 'script-src-attr'.


Harness Error (TIMEOUT), message = null

NOTRUN Should fire a security policy violation for the inline block
FAIL The inline style should not be applied and the attribute style should be applied assert_equals: expected 0 but got 1
